refactor: 정리하지 않은 아이디어에 pin 수정 modal 리팩토링
